requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
langchain==0.1.0
langchain-community==0.0.10
langgraph==0.0.20
tavily-python==0.3.1
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0
asyncio-mqtt==0.16.1 